+++ src/main.rs
@@ -5,12 +5,18 @@ use url::Url;
 use himalaya::{
     backends::{imap_arg, imap_handler, Backend, ImapBackend, MaildirBackend},
     compl::{compl_arg, compl_handler},
-    config::{account_args, config_args, AccountConfig, BackendConfig, DeserializedConfig},
+    config::{
+        account_args, config_args, AccountConfig, BackendConfig, DeserializedConfig,
+        DEFAULT_INBOX_FOLDER,
+    },
     mbox::{mbox_arg, mbox_handler},
     msg::{flag_arg, flag_handler, msg_arg, msg_handler, tpl_arg, tpl_handler},
     output::{output_arg, OutputFmt, StdoutPrinter},
     smtp::LettreService,
 };
+
+#[cfg(feature = "notmuch")]
+use himalaya::{backends::NotmuchBackend, config::MaildirBackendConfig};
 
 fn create_app<'a>() -> clap::App<'a, 'a> {
     clap::App::new(env!("CARGO_PKG_NAME"))
@@ -45,6 +51,10 @@ fn main() -> Result<()> {
 
         let mut imap;
         let mut maildir;
+        #[cfg(feature = "notmuch")]
+        let maildir_config: MaildirBackendConfig;
+        #[cfg(feature = "notmuch")]
+        let mut notmuch;
         let backend: Box<&mut dyn Backend> = match backend_config {
             BackendConfig::Imap(ref imap_config) => {
                 imap = ImapBackend::new(&account_config, imap_config);
@@ -53,6 +63,15 @@ fn main() -> Result<()> {
             BackendConfig::Maildir(ref maildir_config) => {
                 maildir = MaildirBackend::new(&account_config, maildir_config);
                 Box::new(&mut maildir)
+            }
+            #[cfg(feature = "notmuch")]
+            BackendConfig::Notmuch(ref notmuch_config) => {
+                maildir_config = MaildirBackendConfig {
+                    maildir_dir: notmuch_config.notmuch_database_dir.clone(),
+                };
+                maildir = MaildirBackend::new(&account_config, &maildir_config);
+                notmuch = NotmuchBackend::new(&account_config, notmuch_config, &mut maildir)?;
+                Box::new(&mut notmuch)
             }
         };
 
@@ -77,10 +96,15 @@ fn main() -> Result<()> {
         AccountConfig::from_config_and_opt_account_name(&config, m.value_of("account"))?;
     let mbox = m
         .value_of("mbox-source")
-        .unwrap_or(&account_config.inbox_folder);
+        .or_else(|| account_config.mailboxes.get("inbox").map(|s| s.as_str()))
+        .unwrap_or(DEFAULT_INBOX_FOLDER);
     let mut printer = StdoutPrinter::try_from(m.value_of("output"))?;
     let mut imap;
     let mut maildir;
+    #[cfg(feature = "notmuch")]
+    let maildir_config: MaildirBackendConfig;
+    #[cfg(feature = "notmuch")]
+    let mut notmuch;
     let backend: Box<&mut dyn Backend> = match backend_config {
         BackendConfig::Imap(ref imap_config) => {
             imap = ImapBackend::new(&account_config, imap_config);
@@ -89,6 +113,15 @@ fn main() -> Result<()> {
         BackendConfig::Maildir(ref maildir_config) => {
             maildir = MaildirBackend::new(&account_config, maildir_config);
             Box::new(&mut maildir)
+        }
+        #[cfg(feature = "notmuch")]
+        BackendConfig::Notmuch(ref notmuch_config) => {
+            maildir_config = MaildirBackendConfig {
+                maildir_dir: notmuch_config.notmuch_database_dir.clone(),
+            };
+            maildir = MaildirBackend::new(&account_config, &maildir_config);
+            notmuch = NotmuchBackend::new(&account_config, notmuch_config, &mut maildir)?;
+            Box::new(&mut notmuch)
         }
     };